requirements.txt
pandas
openpyxl
reportlab
arabic-reshaper
python-bidi
pyinstaller